+++ solve/src/main.rs
@@ -23,7 +23,7 @@ fn main() {
                 .long("max")
                 .value_name("MAX")
                 .help("Maximum score")
-                .value_parser(clap::value_parser!(u16))
+                .value_parser(clap::value_parser!(u32))
                 .default_value("100"),
         )
         .arg(
@@ -32,7 +32,7 @@ fn main() {
                 .long("sides")
                 .value_name("SIDES")
                 .help("Number of sides on each die")
-                .value_parser(clap::value_parser!(u16))
+                .value_parser(clap::value_parser!(u32))
                 .default_value("6"),
         )
         .arg(
@@ -55,8 +55,8 @@ fn main() {
 
     let args = cli.get_matches();
 
-    let max = *args.get_one::<u16>("max").unwrap();
-    let sides = *args.get_one::<u16>("sides").unwrap();
+    let max = *args.get_one::<u32>("max").unwrap();
+    let sides = *args.get_one::<u32>("sides").unwrap();
     let method = args.get_one::<String>("method").unwrap().as_str();
     let format = args.get_one::<String>("format").unwrap().as_str();
